+++ crates/primitives/src/lib.rs
@@ -55,7 +55,8 @@ pub use signed::{BigIntConversionError, ParseSignedError, Sign, Signed};
 
 mod signature;
 pub use signature::{
-    normalize_v, to_eip155_v, Parity, PrimitiveSignature, Signature, SignatureError,
+    normalize_v, to_eip155_v, ArbitrarySuperSig, EcdsaSignature, K256SuperSig, Parity,
+    PrimitiveSignature, RlpSuperSig, SerdeSuperSig, Signature, SignatureError,
 };
 
 pub mod utils;
